+++ src/main/java/net/raisetech/raisetechtask9/mapper/MovieMapper.java
@@ -1,6 +1,7 @@
-package net.raisetech.raisetechtask9;
+package net.raisetech.raisetechtask9.mapper;
 
 
+import net.raisetech.raisetechtask9.entity.Movie;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 import java.util.List;
