+++ src/main/java/net/raisetech/raisetechtask9/service/MovieService.java
@@ -1,0 +1,11 @@
+package net.raisetech.raisetechtask9.service;
+
+import net.raisetech.raisetechtask9.entity.Movie;
+
+import java.util.List;
+
+public interface MovieService {
+
+    List<Movie> findAll();
+
+}
